split kafka into producer and consumer files
